# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/ai/my-bus": {
      "post": {
        "operationId": "AiController_logMyBusLocation",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiStatusResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "AI"
        ]
      }
    },
    "/api/auth/sign-up": {
      "post": {
        "description": "이메일과 비밀번호로 회원가입",
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpResponseDto"
                }
              }
            }
          }
        },
        "summary": "회원가입",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/sign-in": {
      "post": {
        "description": "이메일과 비밀번호로 로그인",
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponseDto"
                }
              }
            }
          }
        },
        "summary": "로그인",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "description": "인증된 사용자의 정보를 조회",
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "현재 사용자 정보 조회",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "현재 세션 로그아웃",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "로그아웃",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/profiles/me": {
      "get": {
        "description": "현재 사용자의 프로필 정보 조회",
        "operationId": "ProfilesController_getMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "프로필이 없습니다."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 프로필 조회",
        "tags": [
          "Profiles"
        ]
      },
      "put": {
        "description": "현재 사용자의 프로필 정보 수정",
        "operationId": "ProfilesController_updateMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 프로필 수정",
        "tags": [
          "Profiles"
        ]
      }
    },
...
    "/api/jlpt-voca/search": {
      "get": {
        "description": "레벨별 단어 목록 조회 및 검색 (search, word, meaning, meaning_ko)",
        "operationId": "JlptVocaController_list",
        "parameters": [
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "word",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "meaning",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "meaning_ko",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "onlyMeta",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JlptVocaListResponseDto"
                }
              }
            }
          }
        },
        "summary": "JLPT 단어 목록 조회",
        "tags": [
          "JLPT Vocabulary"
        ]
      }
    },
...
    "/api/jlpt-voca/random/level": {
      "get": {
        "description": "레벨 별 또는 전체에서 지정한 개수만큼 랜덤 단어 조회",
        "operationId": "JlptVocaController_randomByLevel",
        "parameters": [
          {
            "name": "level",
            "required": false,
            "in": "query",
            "description": "JLPT 레벨 (미지정 시 전체 레벨에서 조회)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "count",
            "required": false,
            "in": "query",
            "description": "조회할 단어 개수 (최대 50)",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 1,
              "example": 3,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JlptVocaResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "레벨 기반 랜덤 단어 조회",
        "tags": [
          "JLPT Vocabulary"
        ]
      }
    },
    "/api/jlpt-voca/random/ids": {
      "get": {
        "description": "쿼리로 전달된 ID 배열에 해당하는 단어들을 랜덤으로 섞어서 반환",
        "operationId": "JlptVocaController_randomByIds",
        "parameters": [
          {
            "name": "ids",
            "required": false,
            "in": "query",
            "description": "조회할 단어 ID 배열",
            "schema": {
              "example": [
                1,
                2,
                3
              ],
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JlptVocaResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "ID 배열 기반 단어 랜덤 섞기",
        "tags": [
          "JLPT Vocabulary"
        ]
      }
    },
    "/api/jlpt-voca/find": {
      "get": {
        "description": "word 필드가 정확히 일치하는 단어 조회",
        "operationId": "JlptVocaController_findByWordExact",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JlptVocaResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "일치하는 단어가 없습니다."
          }
        },
        "summary": "단어 정확히 일치 검색",
        "tags": [
          "JLPT Vocabulary"
        ]
      }
    },
    "/api/jlpt-voca/examples": {
      "post": {
        "description": "일본어 단어를 입력받아 OpenAI 에이전트를 통해 예문과 한자별 후리가나 매핑을 생성합니다.",
        "operationId": "JlptVocaController_createExampleSentence",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateExampleSentenceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleSentenceResponseDto"
                }
              }
            }
          }
        },
        "summary": "OpenAI 예문 및 후리가나 생성",
        "tags": [
          "JLPT Vocabulary"
        ]
      }
    },
...
  },
  "info": {
    "title": "JLPT Vocabulary API",
    "description": "JLPT 단어 학습을 위한 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AiLocationDto": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "example": 37.5665
          },
          "lng": {
            "type": "number",
            "example": 126.978
          }
        }
      },
      "AiStatusResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "received"
          }
        },
        "required": [
          "status"
        ]
      },
      "SignUpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "7a5a3df3-b70a-47fb-9672-3eb1cc6a92db"
          },
          "email": {
            "type": "string",
            "example": "learner@example.com",
            "format": "email"
          },
          "role": {
            "type": "string",
            "example": "authenticated"
          },
          "aud": {
            "type": "string",
            "example": "authenticated"
          },
          "app_metadata": {
            "type": "object",
            "additionalProperties": true,
            "example": {}
          },
          "user_metadata": {
            "type": "object",
            "additionalProperties": true,
            "example": {}
          },
          "created_at": {
            "type": "string",
            "example": "2026-09-03T12:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "role",
          "aud",
          "app_metadata",
          "user_metadata",
          "created_at"
        ]
      },
      "AuthSessionResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Supabase access token"
          },
          "refresh_token": {
            "type": "string",
            "description": "Supabase refresh token"
          },
          "token_type": {
            "type": "string",
            "example": "bearer"
          },
          "expires_in": {
            "type": "number",
            "example": 3600
          },
          "expires_at": {
            "type": "number",
            "example": 1788436800
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in",
          "user"
        ]
      },
      "SignUpResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthUserResponseDto"
              }
            ]
          },
          "session": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthSessionResponseDto"
              }
            ]
          },
          "message": {
            "type": "string",
            "example": "회원가입이 완료되었습니다."
          }
        },
        "required": [
          "user",
          "session",
          "message"
        ]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SignInResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          },
          "session": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthSessionResponseDto"
              }
            ]
          },
          "access_token": {
            "type": "string",
            "nullable": true
          },
          "refresh_token": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "example": "로그인이 완료되었습니다."
          }
        },
        "required": [
          "user",
          "session",
          "access_token",
          "refresh_token",
          "message"
        ]
      },
      "CurrentUserResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          },
          "message": {
            "type": "string",
            "example": "현재 사용자 정보를 조회했습니다."
          }
        },
        "required": [
          "user",
          "message"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "로그아웃이 완료되었습니다."
          }
        },
        "required": [
          "message"
        ]
      },
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "7a5a3df3-b70a-47fb-9672-3eb1cc6a92db"
          },
          "username": {
            "type": "string",
            "nullable": true,
            "example": "dadok"
          },
          "fullName": {
            "type": "string",
            "nullable": true,
            "example": "다독이"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/avatar.png"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "example": "2026-09-03T12:00:00.000Z",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true,
            "example": "2026-09-03T12:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "username",
          "fullName",
          "avatarUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 80
          },
          "fullName": {
            "type": "string",
            "maxLength": 120
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "JlptVocaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "word": {
            "type": "string",
            "example": "学校"
          },
          "meaning": {
            "type": "string",
            "example": "school"
          },
          "furigana": {
            "type": "string",
            "example": "がっこう"
          },
          "romaji": {
            "type": "string",
            "example": "gakkou"
          },
          "level": {
            "type": "string",
            "example": "5",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5"
            ]
          },
          "meaning_ko": {
            "type": "string",
            "example": "학교"
          }
        },
        "required": [
          "id",
          "word",
          "meaning",
          "furigana",
          "romaji",
          "level",
          "meaning_ko"
        ]
      },
      "JlptVocaPaginationMetaResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 680
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 50
          },
          "totalPages": {
            "type": "number",
            "example": 14,
            "description": "전체 데이터가 0건이면 생략됩니다."
          }
        },
        "required": [
          "total",
          "page",
          "limit"
        ]
      },
      "JlptVocaListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JlptVocaResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/JlptVocaPaginationMetaResponseDto"
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
...
      "GenerateExampleSentenceDto": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string",
            "maxLength": 50,
            "description": "예문을 생성할 일본어 단어",
            "example": "学校"
          }
        },
        "required": [
          "word"
        ]
      },
      "FuriganaPositionResponseDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "number",
            "example": 0,
            "description": "문장 내 한자의 시작 인덱스"
          },
          "end": {
            "type": "number",
            "example": 1,
            "description": "문장 내 한자의 끝 인덱스"
          },
          "text": {
            "type": "string",
            "example": "がく",
            "description": "한자의 후리가나"
          },
          "kanji": {
            "type": "string",
            "example": "学"
          }
        },
        "required": [
          "start",
          "end",
          "text",
          "kanji"
        ]
      },
      "ExampleSentenceResponseDto": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string",
            "example": "学校"
          },
          "word_korean_meaning": {
            "type": "string",
            "example": "학교"
          },
          "sentence": {
            "type": "string",
            "example": "毎日、学校へ行きます。"
          },
          "korean_meaning": {
            "type": "string",
            "example": "매일 학교에 갑니다."
          },
          "furigana_positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FuriganaPositionResponseDto"
            }
          }
        },
        "required": [
          "word",
          "word_korean_meaning",
          "sentence",
          "korean_meaning",
          "furigana_positions"
        ]
      },
...
    }
  }
}